flask_mariadb_docker/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
mysql-connector-python==8.0.26
